backend/requirements.txt
django==5.*
djangorestframework==3.*
psycopg2-binary==2.*
django-cors-headers==4.*
python-dotenv==1.*
requests==2.32.3